[setting] 나머지 모두 패키지 구조 정리..(중간 커밋 까먹음,,)
