+++ app/src/main/java/com/ssu/bilda/presentation/evaluate/ProfileReferFragment.kt
@@ -1,22 +1,18 @@
-package com.ssu.bilda
+package com.ssu.bilda.presentation.evaluate
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
+import androidx.fragment.app.Fragment
+import com.ssu.bilda.R
 
 /**
  * A simple [Fragment] subclass.
- * Use the [MyInfoFragment.newInstance] factory method to
+ * Use the [ProfileReferFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class MyInfoFragment : Fragment() {
+class ProfileReferFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -24,8 +20,8 @@ class MyInfoFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+            param1 = it.getString(param1)
+            param2 = it.getString(param2)
         }
     }
 
@@ -34,7 +30,7 @@ class MyInfoFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_my_info, container, false)
+        return inflater.inflate(R.layout.fragment_profile_refer, container, false)
     }
 
     companion object {
@@ -44,15 +40,15 @@ class MyInfoFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment MyInfoFragment.
+         * @return A new instance of fragment ProfileReferFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            MyInfoFragment().apply {
+            ProfileReferFragment().apply {
                 arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+                    putString(param1, param1)
+                    putString(param2, param2)
                 }
             }
     }
