+++ app/src/main/java/com/ssu/bilda/presentation/evaluate/ProjectStatusFragment.kt
@@ -1,3 +1,5 @@
+package com.ssu.bilda.presentation.evaluate
+
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -6,31 +8,32 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.ssu.bilda.R
-import com.ssu.bilda.presentation.adapter.TeammateNameAdapter
+import com.ssu.bilda.presentation.adapter.SubjectAdapter
 
-class SubjectStatusFragment : Fragment() {
+class ProjectStatusFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // fragment layout inflate
-        val view = inflater.inflate(R.layout.fragment_subject_status, container, false)
+        val view = inflater.inflate(R.layout.fragment_project_status, container, false)
 
         // recyclerview 더미 데이터
-        val teammateNameList = listOf("팀원 1", "팀원 2", "팀원 3")
+        val subjectList = listOf("과목 1", "과목 2", "과목 3")
 
         // adapter 생성, 설정
-        val adapter = TeammateNameAdapter(requireContext(), teammateNameList)
+        val adapter = SubjectAdapter(requireContext(), subjectList)
 
         // RecyclerView를 초기화, adapter 설정
-        val rvTeammateNameList: RecyclerView = view.findViewById(R.id.rv_teammate_name_list)
-        rvTeammateNameList.layoutManager = LinearLayoutManager(requireContext())
-        rvTeammateNameList.adapter = adapter
+        val rvSubjectList: RecyclerView = view.findViewById(R.id.rv_subject_list)
+        rvSubjectList.layoutManager = LinearLayoutManager(requireContext())
+        rvSubjectList.adapter = adapter
 
+        // SubjectAdapter의 클릭 리스너 설정
         adapter.setOnItemClickListener { view ->
             // 아이템이 클릭되었을 때 수행할 동작 구현
-            replaceFragment(TeammateEvalutionFragment())
+            replaceFragment(SubjectStatusFragment())
         }
         return view
     }
