+++ app/src/main/java/com/ssu/bilda/presentation/evaluate/TeammateEvalutionFragment.kt
@@ -1,10 +1,11 @@
+package com.ssu.bilda.presentation.evaluate
+
 import android.app.AlertDialog
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.FrameLayout
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import com.ssu.bilda.R
 import github.hongbeomi.dividerseekbar.DividerSeekBar
