+++ app/src/main/java/com/ssu/bilda/presentation/mypage/ResetPasswordFragment.kt
@@ -1,15 +1,11 @@
-package com.ssu.bilda
+package com.ssu.bilda.presentation.mypage
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
+import androidx.fragment.app.Fragment
+import com.ssu.bilda.R
 
 /**
  * A simple [Fragment] subclass.
@@ -24,8 +20,8 @@ class ResetPasswordFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+            param1 = it.getString(param1)
+            param2 = it.getString(param2)
         }
     }
 
@@ -51,8 +47,8 @@ class ResetPasswordFragment : Fragment() {
         fun newInstance(param1: String, param2: String) =
             ResetPasswordFragment().apply {
                 arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+                    putString(param1, param1)
+                    putString(param2, param2)
                 }
             }
     }
