+++ app/src/main/res/layout/fragment_my_info.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ResetPasswordFragment">
+    tools:context=".presentation.mypage.ResetPasswordFragment">
 
     <LinearLayout
         android:id="@+id/ll_ic_arrow_and_tv_my_info"
