+++ app/src/main/res/layout/fragment_profile_refer.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfileReferFragment">
+    tools:context=".presentation.evaluate.ProfileReferFragment">
 
     <ScrollView
         android:layout_width="0dp"
